Add files via upload
Investigate a Dataset - TMDb Movie Data Analysis.

In this project, we'll be analyzing a dataset and then communicate findings about it. we'll be using the Python libraries NumPy, pandas, and Matplotlib to make your analysis easier.

This data set contains information about 10,000 movies collected from The Movie Database (TMDb), including user ratings and revenue.

Certain columns, like ‘cast’ and ‘genres’, contain multiple values separated by pipe (|) characters.
The final two columns ending with “_adj” show the budget and revenue of the associated movie in terms of 2010 dollars, accounting for inflation over time
There are some odd characters in the ‘cast’ column. They'll be addressed during the data cleaning process

Across the years we have seen movies get released with some being highly anticipated making them do really well in terms of revenue and popularity.
​
This movie data set covers most of the variables that can possibly influence the perfomance of a movie. Take the variable cast for example, there are people who will watch a movie based on the lead cast because maybe they love how he or she acts.
​
The year can also be an influence as different sources have it that in the year 2020, when the world was affected by Corona Virus, movies got more watchtime. 
​
The key questions that this data set can answer are :
​
 *what properties are associated with movies that have high revenue
 
 *what genres are popular from year to year